+++ packages/unigraph-dev-electron/src/main.js
@@ -1,9 +1,12 @@
 const { app, BrowserWindow, Menu, Tray, nativeImage, globalShortcut, shell } = require('electron');
 const path = require('path');
-const { spawn } = require('child_process');
+const os = require('os');
+const child_process = require('child_process');
+const { spawn } = child_process;
 const { fixPathForAsarUnpack } = require('electron-util');
 const { ipcMain } = require('electron');
 const { autoUpdater } = require('electron-updater');
+const log = require('electron-log');
 
 const Store = require('electron-store');
 const net = require('net');
@@ -39,6 +42,13 @@ function isDev() {
     return process.argv[2] == '--dev';
 }
 
+const gotTheLock = app.requestSingleInstanceLock();
+
+if (!gotTheLock) {
+    app.quit();
+    process.exit(0);
+}
+
 if (!isDev()) autoUpdater.checkForUpdatesAndNotify();
 
 function isUnigraphPortOpen(port) {
@@ -67,9 +77,10 @@ async function startServer(logHandler) {
         path.join(userData, 'w'),
         store.get('startServer') !== false,
     );
-    if (store.get('startServer') !== false && !portopen && shouldStartBackend) {
+    if (store.get('startServer') !== false && (!portopen || !isDev()) && shouldStartBackend) {
         const oldConsoleLog = console.log;
         console.log = (data) => {
+            log.log(data);
             if (!Array.isArray(data)) data = [data];
             logs.push(...data);
             if (!dontCheck) checkIfUComplete(...data);
@@ -153,13 +164,6 @@ function createLoadingWindow(props) {
 
     return win;
 }
-
-const createTodayWindow = () =>
-    createMainWindow({
-        transparent: true,
-        frame: false,
-        backgroundColor: '#00ffffff',
-    });
 
 const createOmnibar = () =>
     createMainWindow({
@@ -248,6 +252,12 @@ app.whenReady().then(() => {
                 mainWindow.show();
             }
         });
+        app.on('second-instance', (event, commandLine, workingDirectory) => {
+            // Someone tried to run a second instance, we should focus our window.
+            if (mainWindow) {
+                mainWindow.show();
+            }
+        });
         const windows = [mainWindow, omnibar];
         windows.map((el) =>
             el.on('close', (event) => {
@@ -307,13 +317,22 @@ function showOmnibar() {
 
 ipcMain.on('close_omnibar', closeOmnibar);
 
-app.on('window-all-closed', () => {
-    // Prevent app quitting
+app.on('window-all-closed', (e) => {
+    // Prevent app quitting, instead runs in background
+    e.preventDefault();
 });
 
 app.on('quit', () => {
-    if (alpha) alpha.kill();
-    if (zero) zero.kill();
+    function killer(ps) {
+        if (os.platform() === 'win32'){
+            child_process.exec('taskkill /pid ' + ps.pid + ' /T')
+        } else {
+            ps.kill();  
+        }
+    }
+
+    if (alpha) killer(alpha);
+    if (zero) killer(zero);
 });
 
 app.on('before-quit', function () {