+++ packages/default-packages/unigraph.reddit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/unigraph-dev/unigraph-dev/compare/v0.2.2...v0.2.3) (2022-03-11)
+
+**Note:** Version bump only for package unigraph.reddit
+
+
+
+
+
+## [0.2.2](https://github.com/unigraph-dev/unigraph-dev/compare/v0.2.1...v0.2.2) (2022-03-09)
+
+**Note:** Version bump only for package unigraph.reddit
+
+
+
+
+
 ## [0.2.1](https://github.com/unigraph-dev/unigraph-dev/compare/v0.2.0...v0.2.1) (2022-03-09)
 
 **Note:** Version bump only for package unigraph.reddit